fix: mock-data파일 분리
